refactor: moved to auth context errors
