+++ src/main/java/com/gamix/exceptions/authentication/PasswordUserNotFound.java
@@ -1,4 +1,4 @@
-package com.gamix.exceptions.passwordUser;
+package com.gamix.exceptions.authentication;
 
 import com.gamix.enums.ExceptionMessage;
 import com.gamix.exceptions.ExceptionBase;
